Renames ScanBlockButton to ExplorerButton
