+++ src/ethereum/index.ts
@@ -1,2 +1,2 @@
 export { default as EthHashInfo } from './EthHashInfo';
-export { default as ScanBlockButton } from './ScanBlockButton';
+export { default as ExplorerButton } from './ExplorerButton';
